Test for skip in separate file
