+++ test/test.js
@@ -23,34 +23,4 @@ describe('Array #each', function () {
 		assert.isTrue(callback.withArgs(4,3).calledOnce);
 		assert.isTrue(callback.withArgs(5,4).calledOnce);
 	});
-
-	describe('#skip', function() {
-		it('Should return a new array without the first 2 elements', function() {
-			//Setup
-			var children = [
-		    { name: 'ana', sex: 'f' },
-		    { name: 'fosto', sex: 'm' },
-		    { name: 'jane', sex: 'f' }
-			];
-			var expected = [
-				{ name: 'jane', sex: 'f' }
-			];
-
-			//Execute
-			var response = children.skip(2);
-
-			//Compare
-			assert.deepEqual(response, expected)
-		});
-
-		it('Should throw error. invalid parameter type', function() {
-			//Setup
-			function callback() {
-				[1, 3].skip('string');
-			}
-
-			//Execute and compare
-			assert.throws(callback, TypeError);
-		});
-	});
 });
